add VirtualLock/VirtualUnlock to _pResults and _pDecoys
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@302 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ branches/release_2013023/CometSearch/CometPreprocess.cpp
@@ -264,7 +264,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
       return false;
    }
 
-   pScoring->dwFastXcorrDataSize = pScoring->_spectrumInfoInternal.iArraySize * sizeof(float);
+   pScoring->dwFastXcorrDataSize = (size_t)pScoring->_spectrumInfoInternal.iArraySize * (size_t)sizeof(float);
    pScoring->pfFastXcorrData = (float *)g_cometMemMgr.CometMemAlloc((size_t)pScoring->_spectrumInfoInternal.iArraySize, (size_t)sizeof(float));
    if (NULL == pScoring->pfFastXcorrData)
    {
@@ -285,7 +285,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
             || g_staticParams.ionInformation.iIonVal[ION_SERIES_B]
             || g_staticParams.ionInformation.iIonVal[ION_SERIES_Y]))
    {
-      pScoring->dwFastXcorrDataNLSize = pScoring->_spectrumInfoInternal.iArraySize * sizeof(float);
+      pScoring->dwFastXcorrDataNLSize = (size_t)pScoring->_spectrumInfoInternal.iArraySize * (size_t)sizeof(float);
       pScoring->pfFastXcorrDataNL = (float *)g_cometMemMgr.CometMemAlloc((size_t)pScoring->_spectrumInfoInternal.iArraySize, (size_t)sizeof(float));
       if (NULL == pScoring->pfFastXcorrDataNL)
       {
@@ -368,13 +368,12 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
          }
 
          //MH: Count number of sparse entries needed
-         if(g_staticParams.options.bSparseMatrix && i>0 && !isEqual(pScoring->pfFastXcorrDataNL[i], pScoring->pfFastXcorrDataNL[i-1]))
+         if (g_staticParams.options.bSparseMatrix && i>0 && !isEqual(pScoring->pfFastXcorrDataNL[i], pScoring->pfFastXcorrDataNL[i-1]))
             pScoring->iFastXcorrDataNL++;
       }
    }
 
    free(pPre.pdCorrelationData);
-
    free(pdTmpFastXcorrData);
 
    // Using sparse matrix which means we free pScoring->pfFastXcorrData, ->pfFastXcorrDataNL here
@@ -385,7 +384,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
       pScoring->iFastXcorrDataNL++;
 
       //MH: Fill sparse matrix
-      pScoring->dwSparseFastXcorrDataSize = (size_t)pScoring->iFastXcorrData * sizeof(SparseMatrix);
+      pScoring->dwSparseFastXcorrDataSize = (size_t)pScoring->iFastXcorrData * (size_t)sizeof(SparseMatrix);
       pScoring->pSparseFastXcorrData = (SparseMatrix *)g_cometMemMgr.CometMemAlloc((size_t)pScoring->iFastXcorrData, (size_t)sizeof(SparseMatrix));
       if (pScoring->pSparseFastXcorrData == NULL)
       {
@@ -430,7 +429,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
                || g_staticParams.ionInformation.iIonVal[ION_SERIES_B]
                || g_staticParams.ionInformation.iIonVal[ION_SERIES_Y]))
       {
-         pScoring->dwSparseFastXcorrDataNLSize = (size_t)pScoring->iFastXcorrDataNL * sizeof(SparseMatrix);
+         pScoring->dwSparseFastXcorrDataNLSize = (size_t)pScoring->iFastXcorrDataNL * (size_t)sizeof(SparseMatrix);
          pScoring->pSparseFastXcorrDataNL = (SparseMatrix *)g_cometMemMgr.CometMemAlloc((size_t)pScoring->iFastXcorrDataNL, (size_t)sizeof(SparseMatrix));
          if (pScoring->pSparseFastXcorrDataNL == NULL)
          {
@@ -507,7 +506,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
       // MH: Fill sparse matrix for SpScore - Note that we allocate max number of sp ions,
       // but will use less than this amount.
 
-      pScoring->dwSparseSpScoreDataSize = (size_t)NUM_SP_IONS * sizeof(SparseMatrix);
+      pScoring->dwSparseSpScoreDataSize = (size_t)NUM_SP_IONS * (size_t)sizeof(SparseMatrix);
       pScoring->pSparseSpScoreData = (SparseMatrix *)g_cometMemMgr.CometMemAlloc((size_t)NUM_SP_IONS, (size_t)sizeof(SparseMatrix));
       if (pScoring->pSparseSpScoreData == NULL)
       {
@@ -538,7 +537,7 @@ bool CometPreprocess::Preprocess(struct Query *pScoring,
    }
    else
    {
-      pScoring->dwSpScoreDataSize = (size_t)pScoring->_spectrumInfoInternal.iArraySize * sizeof(float);
+      pScoring->dwSpScoreDataSize = (size_t)pScoring->_spectrumInfoInternal.iArraySize * (size_t)sizeof(float);
       pScoring->pfSpScoreData = (float *)g_cometMemMgr.CometMemAlloc((size_t)pScoring->_spectrumInfoInternal.iArraySize, (size_t)sizeof(float));
       if (pScoring->pfSpScoreData == NULL)
       {
@@ -1208,7 +1207,7 @@ bool CometPreprocess::PeakExtract(double *data,
       }
    }
 
-   memcpy(data, pdPeakExtracted, iArraySize*sizeof(double));
+   memcpy(data, pdPeakExtracted, (size_t)iArraySize*sizeof(double));
    free(pdPeakExtracted);
 
    return true;
